requirements.txt
astroid==1.5.3
click==6.7
Flask==0.12.2
Flask-SocketIO==2.8.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pkg-resources==0.0.0
pycrypto==2.6.1
pylint==1.7.1
python-engineio==1.5.4
python-socketio==1.7.5
six==1.10.0
Werkzeug==0.12.2
wrapt==1.10.10
